Add scroll-triggered opacity change in custom.js
This commit adds a jQuery script to the `custom.js` file. This script changes the opacity of elements with the IDs "banner-left" and "banner-right" based on scroll distance from the top of the page. The code only runs if these two elements are present on the page, preventing unnecessary function calls.